testing and minor changes
