+++ .gitignore
@@ -40,3 +40,5 @@ out
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+dist/
