+++ backend/prisma/seed.ts
@@ -173,7 +173,9 @@ async function main() {
   await prisma.event.deleteMany({});
   await prisma.userPreferences.deleteMany({});
   await prisma.profile.deleteMany({});
+  await prisma.permission.deleteMany({});
   await prisma.user.deleteMany({});
+
   for (const u of userData) {
     const user = await prisma.user.create({
       data: u,
