+++ backend/src/events/views.ts
@@ -1,17 +1,18 @@
-import { Router } from "express";
+import { Router, RequestHandler } from "express";
 import eventController from "./controllers";
 import { auth } from "../middleware/auth"
 
 const eventRouter = Router();
 
+// No provision for auth in test environment for now
+if (process.env.NODE_ENV !== 'test') {
+  eventRouter.use(auth as RequestHandler);
+}
+
 /** Event Specific Routes */
-//@ts-ignore
-eventRouter.get("/", auth, eventController.getEvents);
-//@ts-ignore
-eventRouter.get("/upcoming", auth, eventController.getUpcomingEvents);
-//@ts-ignore
-eventRouter.get("/current", auth, eventController.getCurrentEvents);
-//@ts-ignore
-eventRouter.get("/past", auth, eventController.getPastEvents);
+eventRouter.get("/", eventController.getEvents);
+eventRouter.get("/upcoming", eventController.getUpcomingEvents);
+eventRouter.get("/current", eventController.getCurrentEvents);
+eventRouter.get("/past", eventController.getPastEvents);
 
 export default eventRouter;
