+++ backend/src/server.ts
@@ -4,7 +4,7 @@ import userRouter from "./users/views";
 import eventRouter from "./events/views";
 import swaggerUI from "swagger-ui-express";
 import spec from "../api-spec.json";
-import admin from "firebase-admin"
+
 
 const app: Application = express();
 
@@ -31,27 +31,5 @@ app.get("*", (req, res) => {
 app.post("/", (req, res) => {
   res.send(req.body);
 });
-
-if(!process.env.PRIVATE_KEY){
-  throw new Error("private key is empty");
-}
-
-const serviceAccount = {
-  type: process.env.TYPE,
-  project_id: process.env.PROJECT_ID,
-  private_key_id: process.env.PRIVATE_KEY_ID,
-  private_key: process.env.PRIVATE_KEY.replace(/\\n/g, '\n'),
-  client_email: process.env.CLIENT_EMAIL,
-  client_id: process.env.CLIENT_ID,
-  auth_uri: process.env.AUTH_URI,
-  token_uri: process.env.TOKEN_URI,
-  auth_provider_x509_cert_url: process.env.AUTH_PROVIDER_x509_CERT_URL,
-  client_x509_cert_url: process.env.CLIENT_x509_CERT_URL,
-};
-
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount as admin.ServiceAccount)
-});
-
 
 export default app;
