+++ backend/src/users/views.ts
@@ -1,21 +1,20 @@
-import { Router } from "express";
+import { Router, RequestHandler } from "express";
 import userController from "./controllers";
 import { auth } from "../middleware/auth"
 const userRouter = Router();
 
 /** User Specific Routes */
 
+// No provision for auth in test environment for now
+if (process.env.NODE_ENV !== 'test') {
+  userRouter.use(auth as RequestHandler);
+}
+
 // This approach is cleaner for us becuase we can easily add middle ware
-// HERE IS A TYPESCRIPT COMPILER ERROR THAT WE DON'T KNOW HOW TO FIX :)
-//@ts-ignore
-userRouter.post("/", auth, userController.createUser);
-//@ts-ignore
-userRouter.delete("/:userID", auth, userController.deleteUser);
-//@ts-ignore
-userRouter.put("/:userID", auth, userController.updateUser);
-//@ts-ignore
-userRouter.get("/", auth, userController.getAllUsers);
-//@ts-ignore
-userRouter.get("/search", auth, userController.getSearchedUser);
+userRouter.post("/", userController.createUser);
+userRouter.delete("/:userID", userController.deleteUser);
+userRouter.put("/:userID", userController.updateUser);
+userRouter.get("/", userController.getAllUsers);
+userRouter.get("/search", userController.getSearchedUser);
 
 export default userRouter;
