--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "firebase-admin": "^11.5.0"
-  }
-}
